// apps/web/src/lib/api/generated/models/criterionDimensionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export type CriterionDimensionInput =
	(typeof CriterionDimensionInput)[keyof typeof CriterionDimensionInput];

export const CriterionDimensionInput = {
	population: 'population',
	intervention: 'intervention',
	comparator: 'comparator',
	outcome: 'outcome',
	design: 'design',
	setting: 'setting',
	language: 'language',
	date: 'date',
	other: 'other'
} as const;