// apps/web/src/lib/api/generated/models/frameworkKindInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export type FrameworkKindInput = (typeof FrameworkKindInput)[keyof typeof FrameworkKindInput];

export const FrameworkKindInput = {
	pico: 'pico',
	picos: 'picos',
	peco: 'peco',
	peo: 'peo',
	pcc: 'pcc',
	spider: 'spider',
	custom: 'custom'
} as const;